move MAB mp code to own file
